+++ okayjack/static/okayjack/js/htmx.ext.okayjack.js
@@ -13,6 +13,7 @@
 	// htmx doesn't process these normally. These are new okayjack ones, or those which htmx will only process when they are received in a header from the server
 	const headerAttrs = [
 		'Block',
+		'Partial',
 		'Do-Nothing',
 		'Fire-After-Receive',
 		'Fire-After-Settle',
@@ -21,23 +22,28 @@
 		'Location',
 		'Redirect',
 		'Refresh',
+		'Alert',
 	]
 
 	htmx.defineExtension('okayjack', {
 		onEvent: function (name, evt) {
 			if (name === 'htmx:configRequest') {
 				function appendHxAttribute(attr) {
-					let attrLower = attr.toLowerCase() // attrLower e.g. 'hx-refresh'
-					let blockEl = htmx.closest(evt.detail.elt, "[" + attrLower + "]") // Find the nearest element with the custom attribute
-					if (blockEl) {
-						let value = blockEl.getAttribute(attrLower)
-						
-						// Special case for refresh.
-						// If a url path isn't included, default to the current page
-						if (attrLower.indexOf('refresh') && ((value == '') || (value.toLowerCase() == 'true'))) {
+					const srcElement = evt.srcElement
+					if (srcElement.hasAttribute(attr)) {
+						let value = srcElement.getAttribute(attr)
+						let attrLower = attr.toLowerCase()
+
+						// If a Refresh attribute doesn't specify what path to use to generate the html, use the current path
+						if (attrLower.includes('refresh')  &&  ((value == '') || (value.toLowerCase() == 'true')) ) {
 							value = window.location.pathname + window.location.search
+
+						// base64 encode alert text so it can be sent in a http header
+						} else if (attrLower.includes('alert')) {
+							const bytes = new TextEncoder().encode(value.replace(/\\n/g, '\n')) // Convert text to UTF-8 bytes (Uint8Array)
+							const binaryString = String.fromCharCode(...bytes) // Convert bytes to binary string
+							value = btoa(binaryString) // Encodes binary string to Base64
 						}
-						
 						evt.detail.headers[attr] = value
 					}
 				}
@@ -67,21 +73,44 @@
 	document.addEventListener("htmx:beforeOnLoad", function (e) {
 		const xhr = e.detail.xhr
 		const doNothing = xhr.getResponseHeader("HX-Do-Nothing")
+		const reswap = xhr.getResponseHeader("HX-Reswap")
 
 		if (doNothing) {
 			e.detail.shouldSwap = false
 		}
 
 		if (xhr.status == 422) {
-			// Process 422 status code responses the same way as 200 responses
+			// Process 422 status code responses the same way as 200 responses...
 			e.detail.isError = false
-			if (!doNothing) {
+
+			// ...except if the user specifically chose a noswap option (e.g. HxAlert is a noswap option)
+			if (!doNothing || (reswap != 'none')) {
 				e.detail.shouldSwap = true
 			}
 
 		} else if ((xhr.status >= 400) && (xhr.status < 500)) {
 			e.stopPropagation() // Tell htmx not to process these requests
 			document.children[0].innerHTML = xhr.response // Swap in body of response instead
+		}
+
+
+	})
+
+	/***
+	 * Display alerts 
+	 * 
+	 * We do this after the response is processed (settled) so, to a user, the UI has been updated and the alert is displayed on top.
+	 * 
+	 * The alert text in the header is encoded on the server so it can be included in http headers (which don't support things like new line characters).
+	 */
+	document.addEventListener("htmx:afterSettle", function (e) {
+		const xhr = e.detail.xhr
+		const encodedText = xhr.getResponseHeader("HX-Alert")
+		if (encodedText) {
+			console.log('encodedText', encodedText)
+			const bytes = Uint8Array.from(atob(encodedText), c => c.charCodeAt(0)) // TODO replace this with fromBase64 in the future
+			const decodedText = new TextDecoder().decode(bytes)
+			alert(decodedText)
 		}
 	})