+++ CHANGELOG.rst
@@ -3,6 +3,26 @@
 Changelog
 #########
 
+4.0 (28 Nov 2025)
+=================
+
+Breaking changes
+----------------
+
+* Attributes no longer cascade. htmx 4 attributes don't cascade/inherit by default. Let me know if you need cascading and I'll add the same cascading modifiers (e.g. `:inherit`) as htmx 4.
+
+New
+---
+
+* ``hx-partial`` support 🥳. Works the same as ``hx-block``. The intention is ``hx-partial`` eventually supercedes ``hx-block`` but, as partials are only in Django 6.0, we'll have two options for the forseeable future. And besides, who wants to rename a bunch of blocks to partial?
+* ``HxAlert`` and ``HxErrorAlert`` classes and ``hx-*-alert`` attributes. A HxAlert will display a JavaScript alert when the response has been processed. They all take a single argument which is the alert text.
+
+Improvements
+------------
+
+* swap style is automatically set to ``none`` if no ``block/partial`` has been included. This is useful when using things like ``hx-alert`` as you don't need to set ``hx-swap='none'``.
+
+
 3.0 (18 Nov 2025)
 =================